+++ src/whatsapp/controllers/rabbitmq.controller.ts
@@ -40,6 +40,9 @@ export class RabbitmqController {
         'CONNECTION_UPDATE',
         'CALL',
         'NEW_JWT_TOKEN',
+        'TYPEBOT_START',
+        'TYPEBOT_CHANGE_STATUS',
+        'CHAMA_AI_ACTION',
       ];
     }