+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# 1.5.1 (homolog)
+
+### Feature
+
+* Added listening_from_me option in Set Typebot
+* Added variables options in Start Typebot
+* Added webhooks for typebot events
+* Added ChamaAI integration
+* Added webhook to send errors
+
+### Fixed
+
+* Fix looping connection messages in chatwoot
+
 # 1.5.0 (2023-08-18 12:47)
 
 ### Feature